+++ plugins/fancyholograms/fh-api/src/main/java/com/fancyinnovations/fancyholograms/api/HologramConfiguration.java
@@ -97,4 +97,26 @@ public interface HologramConfiguration {
      * @return Whether the hologram rotation improvement is enabled.
      */
     boolean isHologramRotationImprovementEnabled();
+
+
+    /**
+     * Returns the interval at which holograms check for text updates.
+     *
+     * @return The hologram update interval in milliseconds.
+     */
+    int getHologramUpdateInterval();
+
+    /**
+     * Returns the interval at which backups are created.
+     *
+     * @return The backup interval in hours.
+     */
+    int getBackupInterval();
+
+    /**
+     * Returns how long backups are kept
+     *
+     * @return The backup retention in days.
+     */
+    int getBackupRetention();
 }